Split actions.php into individual files per view
